REFACTOR: Systematic migration continues - contacts and reports
- Implemented contacts blueprint with comprehensive contact management:
  - Contact listing, creation, and viewing functionality
  - Firm-specific contact filtering and client associations
  - Proper access control and data isolation
- Enhanced views blueprint with time tracking reports:
  - Comprehensive time tracking analytics and filtering
  - Billable hours calculations and project breakdowns
  - User and project-specific reporting capabilities
- Progress: 45 routes migrated total (100 → 55 remaining)
- Maintaining marathon pace with systematic blueprint organization
